Fixed up inference script
